--- src/UserList.tsx
@@ -1,8 +1,0 @@
-const userList= {
-  "users": [
-    { "name": "João", "email": "Joao@exemplo.com" },
-    { "name": "Maria", "email": "Maria@exemplo.com" }
-  ]
-}
-
-export default userList;
