+++ src/services/GraphQLOperations.tsx
@@ -16,6 +16,7 @@ export const USERS_QUERY = gql`
   query Users($offset: Int, $limit: Int) {
     users(pageInfo: { offset: $offset, limit: $limit }) {
       nodes {
+        id
         name
         email
       }
